Entity now fits the UML and collisions should be properly implemented
