+++ src/collision/CollisionChecker.java
@@ -1,0 +1,12 @@
+package collision;
+
+import entity.Entity;
+import java.util.List;
+
+public interface CollisionChecker {
+    /**
+     * checks a set of entities for collision
+     * @param entities the entities to check
+     */
+    public void checkCollisions(List<Entity> entities);
+}
